+++ src/test/java/chess/ChessPieceTest.java
@@ -1,11 +1,12 @@
 package chess;
 
+import chess.domain.ChessPiece;
+import chess.domain.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class ChessPieceTest {
     @ParameterizedTest
