+++ src/test/java/chess/CommandTest.java
@@ -1,5 +1,6 @@
 package chess;
 
+import chess.domain.Command;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
