+++ src/test/java/chess/PositionTest.java
@@ -1,6 +1,6 @@
 package chess;
 
-import chess.Position;
+import chess.domain.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
